[Feature #104] Cache generated Descriptors on PU-level.
